requirements.txt
fastapi==0.135.3
uvicorn==0.44.0              # ASGI server to run FastAPI
sqlalchemy==2.0.49           # ORM
psycopg2-binary==2.9.11      # PostgreSQL driver
alembic==1.18.4              # Database migrations (use this, not create_all)
pyjwt==2.12.1                # JWT token creation & verification
pydantic                     # data varification
argon2-cffi==25.1.0          # Password hashing (argon2)
redis==7.4.0                 # Redis client (account lockout, future caching)
pydantic-settings==2.13.1    # Settings management from .env / .key files
fastapi
